% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rFractionalWishart.R
\name{FractionalWishart}
\alias{FractionalWishart}
\title{Fractional Wishart Helper Function}
\usage{
FractionalWishart(df, Sigma, covariance = FALSE)
}
\arguments{
\item{df}{degrees of freedom}

\item{Sigma}{Covariance matrix}

\item{covariance}{should covariance matrix be generated (X / df)}
}
\value{
list or array of random wishart matrices
}
\description{
Fractional Wishart Helper Function
}
\examples{
FractionalWishart(22.5, diag(1, 20))
}
\keyword{internal}
