% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Haff_shrinkage.R
\name{Haff_shrinkage}
\alias{Haff_shrinkage}
\title{Haff Shrinkage Covariance Estimator}
\usage{
Haff_shrinkage(x, ...)
}
\arguments{
\item{x}{data matrix}

\item{...}{other options}
}
\value{
Haff shrinkage covariance estimator
}
\description{
Haff Shrinkage Covariance Estimator
}
\examples{
Haff_shrinkage(as.matrix(iris[-5]))
}
