% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rPsuedoWishart.R
\name{PsuedoWishart}
\alias{PsuedoWishart}
\title{Psuedo Wishart Helper Function}
\usage{
PsuedoWishart(df, Sigma, covariance = FALSE)
}
\arguments{
\item{df}{degrees of freedom}

\item{Sigma}{Covariance matrix}

\item{covariance}{should covariance matrix be generated (X / df)}
}
\value{
list or array of random wishart matrices
}
\description{
Psuedo Wishart Helper Function
}
\examples{
PsuedoWishart(5, diag(1, 20))
}
\keyword{internal}
