% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rWishart.R
\name{rWishart}
\alias{rWishart}
\title{Random Wishart Distributed Matrices}
\usage{
rWishart(n, df, Sigma, covariance = FALSE, simplify = "array")
}
\arguments{
\item{Sigma}{}
}
\description{
Random Wishart Distributed Matrices
}
