% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSingularWishart.R
\name{SingularWishart}
\alias{SingularWishart}
\title{Singular Wishart Helper Function}
\usage{
SingularWishart(df, Sigma, covariance = FALSE)
}
\arguments{
\item{df}{degrees of freedom}

\item{Sigma}{Covariance matrix}

\item{covariance}{should covariance matrix be generated (X / df)}
}
\value{
list or array of random wishart matrices
}
\description{
Singular Wishart Helper Function
}
\examples{
SingularWishart(5, diag(1, 20))
}
\keyword{internal}
