% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tu.R
\name{tu}
\alias{tu}
\title{Function to be optimized for Haff Shrinkage}
\usage{
tu(S, n, p)
}
\arguments{
\item{S}{Sample covariance matrix}

\item{n}{Samples}

\item{p}{Dimensions}
}
\description{
Function to be optimized for Haff Shrinkage
}
\keyword{internal}
