% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rNonsingularWishart.R
\name{NonsingularWishart}
\alias{NonsingularWishart}
\title{Nonsingular Wishart Helper Function}
\usage{
NonsingularWishart(df, Sigma, covariance = FALSE)
}
\arguments{
\item{df}{degrees of freedom}

\item{Sigma}{Covariance matrix}

\item{covariance}{should covariance matrix be generated (X / df)}
}
\value{
list or array of random wishart matrices
}
\description{
Nonsingular Wishart Helper Function
}
\examples{
NonsingularWishart(20, diag(1,5))
}
\keyword{internal}
