% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wishartTest.R
\name{wishartTest}
\alias{wishartTest}
\title{Test if Matrix is a Wishart Matrix}
\usage{
wishartTest(WishMat, Sigma)
}
\arguments{
\item{WishMat}{Wishart Matrix}

\item{Sigma}{Covariance matrix}
}
\description{
Test if Matrix is a Wishart Matrix
}
\examples{
wishartTest(rWishart(1, 5, diag(1, 20), simplify = FALSE)[[1]], diag(1, 20))
}
